src/server/tasks/minecraft/vab_minecraft_src/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvisbase"
version="0.0.1"
description = "JarvisBase Library"
dependencies = [
    "av==11.0.0",
    "torch==2.2.1",
    "torchvision==0.17.1",
    "rich==13.7.1",
    "pyyaml==6.0.1",
    "gymnasium==0.29.1",
    "hydra-core==1.3.2",
    "gym==0.23.1",
    "attrs==23.2.0",
    "gym3==0.3.3"
]
authors = [
    {name="CraftJarvis", email="craftjarvis@outlook.com"},
]
[options]
package_data = {'jarvis'= ['*.json']}

[tool.setuptools.packages.find]
include = ['jarvis*']

[tool.setuptools.dynamic]
readme = {file = "README.md"}

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
